Red Hat bugzilla # 31525
Ref: https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=31525

Work around to a problem with "special" characters in table of contents

This is not a real fix for the problem - but has been in use for a while.
People have document collections which depend on this fix, so while
it is not perfect, it is probably expedient to include it.