Support for BSD (freeBSD 4.4) and MacOS X
Ref: FreeBSD 4.4 patches + Fink patches

This provides equivalent functionality to the freebsd 4.4 patch-aa and
patch-ab patches.

However, the standard BSD patches simply clobber the build environment - so
if you use these, builds on other platforms break.

The problem is that support for the POSIX locale in openjade requires two
things: setlocale, and reasonable wchar support. FreeBSD 4.4 has setlocale
support but is missing some wchar functions that are required. (towupper
for example).

The best solution would be to rework the POSIX locale support so that
it will work with FreeBSD.

However, I've made a simple hack:

- Left the configure setlocale check in place
- Added a new "wchar" check (which tests for towupper).

This defines an additional build parameters SP_HAVE_WCHAR. This is simply
applied with the same scope as SP_HAVE_SETLOCALE.

This means that to get the locale support you must have an environment
where both these are set.

The result is that all environments that worked before will still work...

I've directly edited configure rather than configure.in, but as it is already
impossible to generate configure from configure.in I don't think this
is too much of a problem.