Include re-worked version of RedHat's size_t patch
Ref: redhat openjade-1.3-21

Redhat's RPM contains a patch to provide a single #define to determine
if size_t is the same as unsigned int.

This patch was only applied if the build architecture was s390 or s390x.
However, I can't really see how this patch has a bearing on the S/390.
And the patch didn't actually work with any GNU C++ compiler from 2.8
up to an including 3.0. The idea is quite good though, so I've reworked
the patch to make it work as it is supposed to.

Anyway, this may or may not have a bearing on S/390 support.