More work on autoconf - introduce elements of automake
Ref: Various bugs on source forge.

The Makefiles/configure script was not as clever as it needed to be.
Various people have reported a variety of problems (install directories
not being created etc).

Incorporate some automake elements into the script... which should make
installation a bit more robust.

Also added an "install-man" target to install the manpages.