Extensible message modules
In order to support Windows applications that use DLLs
that use SP's DLL, where the intervening DLLs also
define messages to be used with the SP message handling
infrastructure, this patch reworks the way in which
MessageFragments identify the module within which their
text resides by replacing the (libModule, appModule,
xModule) enumeration with a pointer to a global, opaque
MessageModule object. This MessageModule object then
contains whatever data is needed to identify sets of
text resources on the host system, e.g. an HINSTANCE
DLL handle under Windows, or a message domain string
for gettext.

Only the implementation of the message handling
infrastructure is changed; use of the patch does
require regeneration of .h and .cxx files from .msg
files and recompilation of everything, its general
usage does not change.

Ref: Peter Newcomb, Epremis Corp. Source Forge patch #530161