Nick Kew
Improvements to http:// support. Allow http redirection.
The User-Agent: header can be specified by the environment variable
	SP_HTTP_USER_AGENT
It defaults to "libosp 1.5" if not set.
The environment variable SP_HTTP_ACCEPT can contain the argument
to the Accept: header (was hard coded as "Accept: */*"). There is
no default, so to obtain the same behaviour do
	export SP_HTTP_ACCEPT="*/*"