Neil Roeth
Build Infrastructure/Localisation Changes

The change is to add versions to the translations so that multiple versions
of libosp, i.e., with different SONAMEs, can exist simultaneously without
stomping on each other's files.

Without this, the Debian package would not upgrade properly.�� Since the next
release will require a new SONAME (recent changes made the ABI different), it
makes sense to include this change in the next release.

I accomplished the versioning by adding a version number to SP_MESSAGE_DOMAIN,
and it made sense to use the library SONAME for that.�� So, I created a new
variable in configure.in called SP_LIBOSP_CUR to be the "current" part of the
SONAME, and used that in SP_MESSAGE_DOMAIN, and also referenced that in
lib/Makefile to to actually set the SONAME.