Jessica Perry Hekman
Add new options (-x preserve-case, -x no-overwrite ) to osx

preserve-case:

Preserve casing as specified in the DTD for element names; attribute names;
attribute values which are token lists.

no-overwrite:

When preserving internal or external entities, multiple output files will be
written.  If this option is specified, do not overwrite existing files; instead,
attempt to append a number to the end of the original filename to create a
unique filename.  Exit with an error if too many (100) such files already exist.